requirements.txt
Flask==2.2.2
urllib3==1.26.12
nest-asyncio==1.5.6
uvicorn==0.19.0
numpy==1.23.4
fastapi==0.86.0
chess==1.9.3
discord.py==2.0.1
bs4==0.0.1
requests==2.28.1
aspose-words==22.10.0
aiohttp==3.8.3
aiofiles==22.1.0
pymailtm
easy-pil
fentoboardimage
